Caching improvements, refactor
