+++ ts/.vscode/launch.json
@@ -321,9 +321,10 @@
         "--runTestsByPath",
         "${jest.testFile}"
       ],
-      "cwd": "${workspaceFolder}/packages/knowpro",
+      //"cwd": "${workspaceFolder}/packages/knowpro",
       //"cwd": "${workspaceFolder}/packages/memory/storage",
       //"cwd": "${workspaceFolder}/packages/memory/conversation",
+      "cwd": "${workspaceFolder}/packages/knowledgeProcessor",
       "console": "integratedTerminal",
       "internalConsoleOptions": "neverOpen"
     },
