+++ ts/packages/aiclient/src/index.ts
@@ -10,3 +10,4 @@ export * from "./auth";
 export * from "./tokenCounter";
 export { getChatModelNames, getChatModelMaxConcurrency } from "./modelResource";
 export * as auth from "./auth";
+export * from "./cachingModels";
